Move navigator to core
