+++ core/src/main/java/knaufdan/android/core/di/CoreModule.kt
@@ -6,6 +6,8 @@ import knaufdan.android.core.audio.AudioService
 import knaufdan.android.core.audio.IAudioService
 import knaufdan.android.core.broadcast.BroadcastService
 import knaufdan.android.core.broadcast.IBroadcastService
+import knaufdan.android.core.navigation.INavigator
+import knaufdan.android.core.navigation.Navigator
 import knaufdan.android.core.notification.INotificationService
 import knaufdan.android.core.notification.NotificationService
 import javax.inject.Singleton
@@ -26,4 +28,8 @@ class CoreModule {
     @Provides
     internal fun provideBroadcastService(broadcastService: BroadcastService): IBroadcastService =
         broadcastService
+
+    @Singleton
+    @Provides
+    internal fun provideNavigator(navigator: Navigator): INavigator = navigator
 }
