+++ core/src/main/java/knaufdan/android/core/navigation/INavigator.kt
@@ -1,0 +1,14 @@
+package knaufdan.android.core.navigation
+
+import android.os.Bundle
+
+interface INavigator {
+
+    fun navigateTo(
+        pageNumber: Int,
+        addToBackStack: Boolean,
+        bundle: Bundle
+    )
+
+    fun backPressed()
+}
